# api/swagger/markdown-api.yml
swagger: "2.0"
info:
  version: "1.0"
  title: "Markdown Notes API"
basePath: "/markdown-notes/v1"
consumes:
  - application/json
schemes:
  - http
  - https
paths:
  /users/me:
    get:
      responses:
        200:
          description: Successful response
          schema:
            $ref: "#/definitions/userType"
  /users/{user_id}:
    get:
      parameters:
        - name: user_id
          in: path
          type: string
          description: The user id to retrieve information for
          required: true
      responses:
        200:
          description: Successful response
          schema:
            $ref: "#/definitions/userType"
        403:
          description: Not allowed
        404:
          description: Not found
    put:
      parameters:
        - name: user_id
          in: path
          type: string
          description: The user id to retrieve information for
          required: true
        - name: body
          in: body
          schema:
            $ref: '#/definitions/userType'
          required: true
          description: Updated user information
        
      responses:
        200:
          description: Successfully updated
          schema:
            $ref: "#/definitions/userType"
        403:
          description: Not allowed
        404:
          description: Not found (cannot update unknown user)

  /users/{user_id}/index:
    get:
      parameters:
        - name: user_id
          in: path
          type: string
          description: The user id to retrieve the notes index for
          required: true
      responses:
        200:
          description: Successful response
          schema:
            $ref: "#/definitions/indexType"
        403:
          description: Not allowed
        404:
          description: Not found
          
  /notes/{note_id}:
    get:
      parameters:
        - name: note_id
          in: path
          type: string
          description: The id of the note to retrieve
          required: true
      responses:
        200:
          description: Successful response
          schema:
            $ref: "#/definitions/noteType"
        403:
          description: Not allowed
        404:
          description: Not found
    put:
      parameters:
        - name: note_id
          in: path
          type: string
          description: The id of the note to retrieve
          required: true
        - name: body
          in: body
          schema:
            $ref: "#/definitions/noteType"
          required: true
          description: The note to update
      responses:
        204:
          description: Successfully saved
        400:
          description: Bad request. User needs registration first.
        403:
          description: Not allowed to store this
    delete:
      parameters:
        - name: note_id
          in: path
          type: string
          description: The id of the note to retrieve
          required: true
      responses:
        204:
          description: Successfully deleted, no content
        403:
          description: Not allowed
        404:
          description: Not found

definitions:
  userType:
    type: object
    properties:
      registered:
        type: boolean
      external_id:
        type: string
      username:
        type: string
      name:
        type: string
      email:
        type: string
      company:
        type: string
      created_date:
        type: number
      updated_date:
        type: number

  indexType:
    type: array
    items:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
    
  noteType:
    type: object
    properties:
      id:
        type: string
      title:
        type: string
      markdown:
        type: string
      owner:
        type: string
      created_date:
        type: number
      created_by:
        type: string
      updated_date:
        type: number
      updated_by:
        type: string